+++ Client/Components/MBSheet/Stepper/MBSheetStepperModel.cs
@@ -5,6 +5,6 @@ public class MBSheetStepperModel
     public string Uom { get; set; }
     public string ServiceNo { get; set; }
     public string ShortServiceDesc { get; set; }
-    public float CumulativeMeasuredQty { get; set; }
-    public float AvailableMeasurementQty { get; set; }
+    public decimal CumulativeMeasuredQty { get; set; }
+    public decimal AvailableMeasurementQty { get; set; }
 }
